+++ R/databases.R
@@ -106,19 +106,23 @@ getDatabaseTree <- function(databaseId) {
 #'
 #' @export
 getDatabaseResources <- function(database) {
-  if(is.character(database)) {
-    databaseTree <- getDatabaseTree(database)  
-  } else if(is.list(database)) {
-    databaseTree <- database
-  } else {
-    stop("The `database` argument must be a database id or a databaseTree")
-  }
+  UseMethod("getDatabaseResources")
+}
+
+#' @export
+getDatabaseResources.character <- function(database) {
+  tree <- getDatabaseTree(database)  
+  getDatabaseResources(tree)
+}
+
+#' @export
+getDatabaseResources.databaseTree <- function(database) {
   dplyr::tibble(
-    id = unlist(lapply(databaseTree$resources, function(x) {x$id})),
-    label = unlist(lapply(databaseTree$resources, function(x) {x$label})),
-    type = unlist(lapply(databaseTree$resources, function(x) {x$type})),
-    parentId = unlist(lapply(databaseTree$resources, function(x) {x$parentId})),
-    visibility = unlist(lapply(databaseTree$resources, function(x) {x$visibility}))
+    id = unlist(lapply(database$resources, function(x) {x$id})),
+    label = unlist(lapply(database$resources, function(x) {x$label})),
+    type = unlist(lapply(database$resources, function(x) {x$type})),
+    parentId = unlist(lapply(database$resources, function(x) {x$parentId})),
+    visibility = unlist(lapply(database$resources, function(x) {x$visibility}))
   )
 }
 
@@ -215,8 +219,11 @@ getDatabaseUsers <- function(databaseId, asDataFrame = TRUE) {
       version = unlist(lapply(users, function(x) {x$version})),
       inviteDate = as.Date(unlist(lapply(users, function(x) {x$inviteDate}))),
       deliveryStatus = unlist(lapply(users, function(x) {x$deliveryStatus})),
-      inviteAccepted = unlist(lapply(users, function(x) {x$inviteAccepted})) # ,
-      # role = lapply(users, function(x) {x$role})
+      inviteAccepted = unlist(lapply(users, function(x) {x$inviteAccepted})),
+      locked = unlist(lapply(users, function(x) {x$locked})),
+      userLicenseType = unlist(lapply(users, function(x) {x$userLicenseType})),
+      lastLoginDate = as.Date(unlist(lapply(users, function(x) {if(is.null(x$lastLoginDate)) NA else x$lastLoginDate}))),
+      activationStatus = unlist(lapply(users, function(x) {x$activationStatus}))
     )
     
     usersDF$role <- lapply(users, function(x) {x$role})
@@ -318,7 +325,7 @@ checkUserRole <- function(databaseId, newUser, roleId, roleParameters, roleResou
 
 #' addDatabaseUser
 #'
-#' Invites a user to a database.
+#' Invites a user to a database and assigns a role
 #'
 #' @param databaseId the id of the database to which they should be added
 #' @param email the user's email
@@ -326,7 +333,8 @@ checkUserRole <- function(databaseId, newUser, roleId, roleParameters, roleResou
 #' @param locale the locale ("en', "fr", "ar", etc) to use inviting the user (only used if they do not already have an ActivityInfo account)
 #' @param roleId the id of the role to assign to the user.
 #' @param roleParameters a named list containing the role parameter values
-#' @param roleResources a list of folders in which this role should be assigned (or the databaseId if they should have this role in the whole database)
+#' @param roleResources an optional list of optional grant-based resources assigned to the user
+#' @param assignment optionally create and pass a \code{\link[activityinfo]{roleAssignment}} like in updateUserRole()
 #'
 #' @details
 #'
@@ -347,6 +355,8 @@ checkUserRole <- function(databaseId, newUser, roleId, roleParameters, roleResou
 #' in many database templates has a `partner` parameter that is used to filter which
 #' records are visible to the user. The value of this parameter is the record id of the
 #' user's partner in the related Partner form.
+#' 
+#' Optional grants can be specified by adding the resource id of those grants to a list and passing that to `roleResources`.
 #'
 #' @examples
 #' \dontrun{
@@ -372,9 +382,17 @@ checkUserRole <- function(databaseId, newUser, roleId, roleParameters, roleResou
 #' @export
 addDatabaseUser <- function(databaseId, email, name, locale = NA_character_, roleId,
                             roleParameters = list(),
-                            roleResources = c(databaseId)) {
+                            roleResources = c(databaseId), assignment) {
 
   url <- paste(activityInfoRootUrl(), "resources", "databases", databaseId, "users", sep = "/")
+  
+  if (!missing(assignment)) {
+    stopifnot("An assignment must be created with roleAssignment()" = ("activityInfoRoleAssignment" %in% class(assignment)))
+    stopifnot("Either an assignment must be provided or roleId to addDatabaseUser(), but not both." = missing(roleId))
+    roleId = assignment$id
+    roleParameters = assignment$parameters
+    roleResources = assignment$resources
+  }
 
   request <- list(
     email = email,
@@ -415,6 +433,44 @@ addDatabaseUser <- function(databaseId, email, name, locale = NA_character_, rol
   }
 }
 
+
+#' getDatabaseRoles
+#'
+#' Get database roles in a data frame.
+#'
+#' @param database database tree using \link{getDatabaseTree} or the databaseId
+#'
+#' @examples
+#' \dontrun{
+#' dbTree <- getDatabaseTree(databaseId = "ck3pqrp9a1z") # fetch the database tree
+#' roles <- getDatabaseRoles(dbTree) # get the database roles
+#' }
+#' @export
+#'
+getDatabaseRoles <- function(database) {
+  UseMethod("getDatabaseRoles")
+}
+
+#' @export
+getDatabaseRoles.character <- function(database) {
+  tree <- getDatabaseTree(databaseId = database)
+  getDatabaseRoles(tree)
+}
+
+#' @export
+getDatabaseRoles.databaseTree <- function(database) {
+  dplyr::tibble(
+    id = unlist(lapply(database$roles, function(x) {x$id})),
+    label = unlist(lapply(database$roles, function(x) {x$label})),
+    permissions = lapply(database$roles, function(x) {x$permissions}),
+    parameters = lapply(database$roles, function(x) {x$parameters}),
+    filters = lapply(database$roles, function(x) {x$filters}),
+    grants = lapply(database$roles, function(x) {x$grants}),
+    version = unlist(lapply(database$roles, function(x) {x$version})),
+    grantBased = unlist(lapply(database$roles, function(x) {x$grantBased}))
+  )
+}
+
 #' getDatabaseRole
 #'
 #' Helper method to fetch a role based on its id using the database tree or database id.
@@ -428,7 +484,6 @@ addDatabaseUser <- function(databaseId, email, name, locale = NA_character_, rol
 #' dbTree <- getDatabaseTree(databaseId = "ck3pqrp9a1z") # fetch the database tree
 #' role <- getDatabaseRole(dbTree, roleId = "rp") # extract the reporting partner role
 #' }
-#' 
 #' @export
 #'
 getDatabaseRole <- function(database, roleId) {
@@ -585,7 +640,7 @@ updateUserRole <- function(databaseId, userId, assignment) {
 roleAssignment <- function(roleId, roleParameters = list(), roleResources) {
   stopifnot(is.list(roleParameters))
   if (any(is.na(names(roleParameters)))) {
-    stop("roleParameters must be named with each parameter name.")
+    stop("In the `roleParameters` list, each item must be named")
   }
 
   if (length(roleParameters) == 0) {
@@ -935,7 +990,8 @@ deleteRoles <- function(databaseId, roleIds) {
 #' See \link{role} for the creation of roles.
 #'
 #' @param id the id of the parameter, for example "partner", which can
-#' be used in a formula as "@user.partner" 
+#' be used in a formula as "@user.partner". The id starts with a letter and may
+#' contain letters, numbers and underscores _ under 32 characters.
 #' @param label the label of the partner, for example, "Reporting partner"
 #' @param range the id of a reference table, for example the list of partners, 
 #' or a formula
@@ -949,7 +1005,7 @@ deleteRoles <- function(databaseId, roleIds) {
 #' }
 parameter <- function(id, label, range) {
   stopifnot("The id must be a character string" = is.null(id)||(is.character(id)&&length(id)==1&&nchar(id)>0))
-  stopifnot("The id must start with a letter, must be made of letters and underscores _ and cannot be longer than 32 characters" = is.null(id)||grepl("^[A-Za-z][A-Za-z0-9_]{0,31}$", id))
+  stopifnot("The id must start with a letter, must be made of letters, numbers and underscores _ and cannot be longer than 32 characters" = is.null(id)||grepl("^[A-Za-z][A-Za-z0-9_]{0,31}$", id))
   stopifnot("The label is required to be a character string" = (is.character(label)&&length(label)==1&&nchar(label)>0))
   stopifnot("The range is required and must be a character string" = !is.null(range)&&(is.character(range)&&length(range)==1&&nchar(range)>0))
   
@@ -1063,7 +1119,8 @@ roleFilter <- function(id, label, filter) {
 #' Some administrative permissions are defined at the level of the role rather 
 #' than within grants. See \link{databasePermissions}.
 #'
-#' @param id the id of the role
+#' @param id the id of the role, must start with a lower case letter and may 
+#' contain up to 32 lower case letters and numbers 
 #' @param label the label or name of the role, e.g. "Viewer" or "Administrator" 
 #' @param parameters a list of \link{parameter} items defining role parameters
 #' @param grants a list of \link{grant} items for each resource and their 
@@ -1103,7 +1160,7 @@ roleFilter <- function(id, label, filter) {
 #' }
 role <- function(id, label, parameters = list(), grants, permissions = databasePermissions()) {
   stopifnot("The id must be a character string" = is.null(id)||(is.character(id)&&length(id)==1&&nchar(id)>0))
-  stopifnot("The id must start with a letter, must be made of lowercase letters and underscores _ and cannot be longer than 32 characters" = is.null(id)||grepl("^[a-z][a-z0-9_]{0,31}$", id))
+  stopifnot("The id must start with a letter, must be made of lowercase letters and numbers and cannot be longer than 32 characters" = is.null(id)||grepl("^[a-z][a-z0-9]{0,31}$", id))
   
   stopifnot("The label is required to be a character string" = (is.character(label)&&length(label)==1&&nchar(label)>0))